+++ libs/nx_utils/src/nx/utils/ipc/named_pipe_socket_unix.cpp
@@ -114,7 +114,7 @@ SystemError::ErrorCode NamedPipeSocket::read(
                     continue;
                 return errCode;
             }
-            *bytesRead = numberOfBytesRead;
+            *bytesRead = (unsigned int) numberOfBytesRead;
             return SystemError::noError;
         }
         else
